requirements.txt
openai
yfinance
python-dotenv
numpy
pandas
scipy

